Make the project pip installable and unpin versions
The project was failing to install into a clean environment.

Airflow is a completely separate complex application. I think it is best
to have it as a separate project using docker compose or K8.

You do not need to pin versions at this prototype stage of development.
Once you have a production version then pin otherwise keep code up to
date for as long as possible.